Test mutual failover setup
* Resolves #67
* Proves expected behavior of primary-primary failover setup
* Move common test function to `helpers.go`

```
--- PASS: TestOhmyglbSplitFailoverExample (230.45s)
    --- PASS: TestOhmyglbSplitFailoverExample/Each_cluster_resolves_its_own_set_of_IP_addresses (30.53s)
    --- PASS: TestOhmyglbSplitFailoverExample/serviceHealth_becomes_Unhealthy_after_scaling_down_to_0 (10.17s)
    --- PASS: TestOhmyglbSplitFailoverExample/Cluster_1_failovers_to_Cluster_2 (61.28s)
    --- PASS: TestOhmyglbSplitFailoverExample/Cluster_2_still_returns_own_entries (0.07s)
```
